requirements.txt
pytorch-cpu
torchvision-cpu
pillow
argparse
